requirements.txt
aiogram==2.21
emoji==1.7.0
matplotlib==3.5.2
numpy==1.22.4
Pillow==9.2.0
requests==2.27.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu